# docs/swagger.yaml
basePath: /notifications/api
definitions:
  CoreUserRef:
    properties:
      name:
        type: string
      user_id:
        type: string
    type: object
...
  Recipient:
    properties:
      name:
        type: string
      notification_disabled:
        type: boolean
      user_id:
        type: string
    type: object
...
  model.Message:
    description: wraps all needed information for the notification
    properties:
      body:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      date_created:
        type: string
      date_updated:
        type: string
      id:
        type: string
      priority:
        type: integer
      recipients:
        items:
          $ref: '#/definitions/Recipient'
        type: array
      recipients_criteria_list:
        items:
          $ref: '#/definitions/model.RecipientCriteria'
        type: array
      sender:
        $ref: '#/definitions/model.Sender'
      subject:
        type: string
      topic:
        type: string
    type: object
  model.RecipientCriteria:
    properties:
      app_platform:
        type: string
      app_version:
        type: string
    type: object
  model.Sender:
    properties:
      type:
        description: user or system
        type: string
      user:
        $ref: '#/definitions/CoreUserRef'
    type: object
  sendMailRequestBody:
    properties:
      body:
        type: string
      subject:
        type: string
      to_mail:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Rokwire Notifications Building Block API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rokwire Notifications Building Block API
  version: 1.1.8
paths:
...
  /admin/message:
    post:
      consumes:
      - application/json
      description: Creates a message
      operationId: CreateMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - AdminUserAuth: []
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates a message
      operationId: UpdateMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - AdminUserAuth: []
      tags:
      - Admin
...
  /int/mail:
    post:
      description: Sends an email
      operationId: InternalSendMail
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sendMailRequestBody'
      produces:
      - text/plain
      responses:
        "200":
          description: ""
      security:
      - InternalAuth: []
      tags:
      - Internal
  /int/message:
    post:
      description: Sends a message to a user, list of users or a topic
      operationId: InternalSendMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - InternalAuth: []
      tags:
      - Internal
  /message:
    post:
      consumes:
      - application/json
      description: Creates a message. Message without subject and body will be interpreted
        as a data massage and it won't be stored in the database
      operationId: createMessage
      parameters:
      - description: body json
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Message'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - UserAuth: []
      tags:
      - Client
...
schemes:
- https
securityDefinitions:
  AdminUserAuth:
    in: header (add admin core access token with Bearer prefix to the Authorization
      value. The token must contain notifications_admin as a permission)
    name: Authorization
    type: apiKey
  InternalAuth:
    in: header
    name: INTERNAL-API-KEY
    type: apiKey
  RokwireAuth:
    in: header
    name: ROKWIRE-API-KEY
    type: apiKey
  UserAuth:
    in: header (add core access token with Bearer prefix to the Authorization value.
      The token must represent either anonymous or authenticated user )
    name: Authorization
    type: apiKey
swagger: "2.0"